+++ Nukepayload2.Csv/Attributes/CsvIgnoreAttribute.vb
@@ -1,0 +1,8 @@
+﻿''' <summary>
+''' Indicates this property won't be serialized or be deserialized.
+''' </summary>
+<AttributeUsage(AttributeTargets.Property)>
+Public Class CsvIgnoreAttribute
+    Inherits Attribute
+
+End Class