+++ Nukepayload2.Csv/Attributes/CsvPropertyAttribute.vb
@@ -1,0 +1,8 @@
+﻿''' <summary>
+''' Indicates this property will be serialized or be deserialized.
+''' </summary>
+<AttributeUsage(AttributeTargets.Property)>
+Public Class CsvPropertyAttribute
+    Inherits Attribute
+
+End Class