+++ Nukepayload2.Csv/CsvSheetInfo.vb
@@ -1,0 +1,4 @@
+﻿Friend Structure CsvSheetInfo
+    Public ColumnInfo As CsvColumnInfo()
+    Public ColumnOrderKind As CsvColumnOrderKind?
+End Structure